requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
python-gitlab==4.4.0
requests==2.31.0
requests-toolbelt==1.0.0
urllib3==2.2.1
